+++ modules/alloy/src/main/ts/ephox/alloy/positioning/mode/NodeAnchor.ts
@@ -1,6 +1,7 @@
 import { FieldSchema } from '@ephox/boulder';
 import { Optional } from '@ephox/katamari';
 import { SugarBody } from '@ephox/sugar';
+import { Bounds } from '../../alien/Boxes';
 
 import { AlloyComponent } from '../../api/component/ComponentApi';
 import * as Fields from '../../data/Fields';
@@ -10,15 +11,16 @@ import * as AnchorLayouts from './AnchorLayouts';
 import * as ContainerOffsets from './ContainerOffsets';
 import * as ContentAnchorCommon from './ContentAnchorCommon';
 
-const placement = (component: AlloyComponent, anchorInfo: NodeAnchor, origin: Origins.OriginAdt): Optional<Anchoring> => {
+const placement = (component: AlloyComponent, anchorInfo: NodeAnchor, origin: Origins.OriginAdt, getBounds: Optional<() => Bounds>): Optional<Anchoring> => {
   const rootPoint = ContainerOffsets.getRootPoint(component, origin, anchorInfo);
 
   return anchorInfo.node
     // Ensure the node is still attached, otherwise we can't get a valid client rect for a detached node
     .filter(SugarBody.inBody)
     .bind((target) => {
+      const cappingViewport = Origins.viewport(origin, getBounds);
       const rect = target.dom.getBoundingClientRect();
-      const nodeBox = ContentAnchorCommon.capRect(rect.left, rect.top, rect.width, rect.height);
+      const nodeBox = ContentAnchorCommon.capRect(rect.left, rect.top, rect.width, rect.height, Optional.some(cappingViewport));
       const elem = anchorInfo.node.getOr(component.element);
       return ContentAnchorCommon.calcNewAnchor(nodeBox, rootPoint, anchorInfo, origin, elem);
     });