+++ modules/alloy/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Added a new `setValue` method to the `SliderUi`. This allows the slider value to be modified without firing a change event.
+- Fixed a case where the context toolbar would not stick to the bottom of the view.
 
 ## 8.1.0 - 2020-11-18